requirements.txt
pytest
python-dateutil
astropy
pandas
numpy
joblib
matplotlib